+++ ostap/fitting/signals.py
@@ -152,6 +152,9 @@ from   ostap.fitting.fithelpers import ( Phases  , ZERO      ,
                                          Tail    , LeftTail  , RightTail ,
                                          SigmaLR , TwoSigmas )    
 from   ostap.fitting.variables  import var_tanh, SETVAR
+## 
+from   ostap.fitting.wide       import BreitWigner_pdf 
+## 
 import ostap.math.dalitz
 import ostap.math.models
 import ROOT, math