+++ ostap/stats/tests/test_stats_gof1d.py
@@ -7,6 +7,12 @@
 # ==============================================================================
 """ Test Goodness-of-fits for 1D fits 
 """
+# ============================================================================== 
+## ATTENTION! 
+import os 
+os.environ [ "OMP_NUM_THREADS"      ]  = "1"
+os.environ [ "MKL_NUM_THREADS"      ]  = "1"
+os.environ [ "OPENBLAS_NUM_THREADS" ]  = "1"
 # ==============================================================================
 from   ostap.plotting.canvas  import use_canvas
 from   ostap.utils.ranges     import vrange
@@ -58,16 +64,40 @@ gauss.sigma.fix ( 1.0 )
 
 fitconf = { 'draw' : True , 'nbins' : 25 , 'refit' : 5 , 'quiet' : True }
 
+small   = numcpu () <= 8
+
+# ==============================================================================
+## Run Mix-Sample Goodness-of-Fit test
+def run_MIX ( pdf, data, result , label , logger = logger ) :
+    """ Run Mixed-Sample Goodness-of-Fit test
+    """
+
+    nToys = 10 if small else 200 
+
+    logger.info ( 'Run Mixed-Sample GoF-test %s' % label )
+
+    with timing ( 'MIX-test %s' %  label  , logger = logger ) :
+        
+        gof = GnD.MIX ( nToys = nToys )
+        pdf.load_params ( result , silent = True ) 
+        tvalue         = gof          ( pdf , data )
+        tvalue, pvalue = gof.pvalue   ( pdf , data , tvalue = tvalue )
+    
+    title  = 'Goodness-of-Fit Mixed-Samples-test %s' % label     
+    table  = gof.report ( title = title , prefix = '# ' )
+    logger.info ( '%s:\n%s' % ( title , table ) )
+
 # ==============================================================================
 ## Run Point-to-Point dissimilatity Goodness-of-Fit test
 def run_PPD ( pdf, data, result , label , logger = logger ) :
     """ Run Point-to-Point dissimilarity Goodness-of-Fit test
     """
 
-    nToys = 10 if numcpu () < 10 else 200 
-    Ns    =  5 if numcpu () < 10 else  10
-    
-    logger.info ( 'Run Point-to-Point Dissimilarity GoF-test %s for %d different values of sigma' % ( label , Ns  ) )
+    nToys = 10 if small else 200     
+    Ns    =  2 if small else   5
+
+    Ns = 1
+    logger.info ( 'Run Point-to-Point Dissimilarity GoF-test %s for %d different values of sigma' % ( label , Ns + 1   ) )
 
     header = ()
     rows   = [] 
@@ -75,14 +105,12 @@ def run_PPD ( pdf, data, result , label , logger = logger ) :
 
         with timing ( 'PPD-test %s sigma=%.3f' %  ( label  , sigma ) , logger = logger ) :
             
-            ppd = GnD.PPD ( nToys = nToys , sigma = sigma )
+            gof = GnD.PPD ( nToys = nToys , sigma = sigma )
             pdf.load_params ( result , silent = True ) 
-            tvalue         = ppd          ( pdf , data )
-            tvalue, pvalue = ppd.pvalue   ( pdf , data )
-            
-            ecdf           = ppd.ecdf
-            
-            header , row   = ppd.the_row ( tvalue = tvalue ,
+            tvalue         = gof          ( pdf , data )
+            tvalue, pvalue = gof.pvalue   ( pdf , data , tvalue = tvalue )            
+            ecdf           = gof.ecdf            
+            header , row   = gof.the_row ( tvalue = tvalue ,
                                            pvalue = pvalue ,
                                            ecdf   = ecdf   )
             
@@ -96,30 +124,27 @@ def run_PPD ( pdf, data, result , label , logger = logger ) :
     table  = T.table ( rows , title = title , prefix = '# ' , alignment = 'lccccccc' )
     logger.info ( '%s:\n%s' % ( title , table ) )
 
+
 # ==============================================================================
 ## Run Distance-to-Nearest-Neighbour Goodness-of-Fit test
 def run_DNN  ( pdf , data , result , label , logger = logger ) :
     """ Run Distance-to-Nearest-Neighbour Goodness-of-Fit test
     """
 
-    nToys = 10 if numcpu () < 10 else 200 
+    nToys = 10 if small else 200     
     
     with timing ( 'DNN-test %s' % label , logger = logger ) :
         
-        dnn = GnD.DNN   ( nToys = nToys , histo = 51 )        
-        pdf.load_params ( result , silent = True )
-        
-        tvalue         = dnn          ( pdf , data )
-        tvalue, pvalue = dnn.pvalue   ( pdf , data )
+        gof            = GnD.DNN   ( nToys = nToys , histo = 51 )        
+        pdf.load_params ( result , silent = True )        
+        tvalue         = gof          ( pdf , data )
+        tvalue, pvalue = gof.pvalue   ( pdf , data , tvalue = tvalue )
 
-    with use_canvas ( 'DNN-toys %s'  % label ) : dnn.draw ( tvalue = tvalue ) 
+    with use_canvas ( 'DNN-toys %s'  % label ) : gof.draw ( tvalue = tvalue ) 
 
     title  = 'Goodness-of-Fit DNN-test %s' % label     
-    table  = dnn.table ( title = title , prefix = '# ' )
+    table  = gof.report ( title = title , prefix = '# ' )
     logger.info ( '%s:\n%s' % ( title , table ) )
-
-    ## u-distribution
-    return dnn.histo 
 
 # ==============================================================================
 ## Run USTAT Goodness-of-Fit test
@@ -127,23 +152,20 @@ def run_USTAT  ( pdf , data, result , label , logger = logger ) :
     """ Run USTAT Goodness-of-Fit test
     """
 
-    nToys = 10 if numcpu () < 10 else 200 
+    nToys = 10 if small else 200     
 
-    with timing ( 'uStat-test' , logger = logger ) : 
-        ustat = USTAT   ( nToys = nToys , histo = 110 , parallel = True  )        
+    with timing ( 'uStat-test %s' % label , logger = logger ) : 
+        gof = USTAT   ( nToys = nToys , histo = 110 , parallel = True  )        
         pdf.load_params ( result , silent = True )
-        tvalue         = ustat        ( pdf , data )    
-        tvalue, pvalue = ustat.pvalue ( pdf , data )
+        tvalue         = gof        ( pdf , data )    
+        tvalue, pvalue = gof.pvalue ( pdf , data , tvalue = tvalue )
         
-    with use_canvas ( 'USTAT-toys %s'  % label ) :
-        ustat.draw ( tvalue = tvalue ) 
+    ##with use_canvas ( 'USTAT-toys %s'  % label ) :
+    ##    ustat.draw ( tvalue = tvalue ) 
         
     title  = 'Goodness-of-Fit USTAT-test %s' % label     
-    table  = ustat.table ( title = title , prefix = '# ' )
+    table  = gof.report ( title = title , prefix = '# ' )
     logger.info ( '%s:\n%s' % ( title , table ) )
-
-    ## u-distribution
-    return ustat.histo 
 
 plots = [] 
 # ==============================================================================
@@ -151,7 +173,10 @@ def run_fit ( pdf , dataset , label  , logger = logger ) :
     """ Make a test for presumably GOOD or BAD fits
     """
 
-    nToys = 100 if numcpu () < 10 else 1000
+    nToys = 10 if small else 200     
+
+    nToys = 20
+
     
     logger.info ( 'Make a test for presumably %s fit' % label  )
 
@@ -171,8 +196,7 @@ def run_fit ( pdf , dataset , label  , logger = logger ) :
             toys = toys.run ( nToys = nToys , parallel = True )
         logger.info ( 'Goodness-of-fit (%s) with %d toys:\n%s' % ( label , toys.nToys , toys ) ) 
 
-    return
-
+    """
     with use_canvas ( title = '%s:GoF/Kolmogorov-Smirnov' % label ) :
         dks = toys.draw ( 'Kolmogorov-Smirnov')
         plots.append ( dks ) 
@@ -205,13 +229,16 @@ def run_fit ( pdf , dataset , label  , logger = logger ) :
         plots.append ( dai )         
     with use_canvas ( title = '%s:GoF/BayesianIC'          % label ) :
         dbi = toys.draw ( 'BIC')
-        plots.append ( dbi )         
+        plots.append ( dbi )
+    
+    """
     
     ## Try to use multidimensional methods:
     
+    ## run_MIX   ( pdf , dataset , r , label , logger )    
     ## run_PPD   ( pdf , dataset , r , label , logger )    
     ## run_DNN   ( pdf , dataset , r , label , logger )
-    ## run_USTAT ( pdf , dataset , r , label , logger )
+    run_USTAT ( pdf , dataset , r , label , logger )
 
 # =====================================================================================
 def test_good_fit_1 ( ) :